(PC-13519)[PRO] refactor: create Offers route and put initial logic in it
