+++ pro/src/routes/Offers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Offers'
